+++ threedigrid/admin/lines/exporters.py
@@ -96,6 +96,8 @@ class LinesOgrExporter(BaseOgrExporter):
         node_a = line_data['line'][0]
         node_b = line_data['line'][1]
         for i in range(node_a.size):
+            if line_data['id'][i] == 0:
+                continue  # skip the dummy element
             if geom_source == 'from_threedicore':
                 line = ogr.Geometry(ogr.wkbLineString)
                 line.AddPoint(line_data['line_coords'][0][i],
@@ -113,22 +115,22 @@ class LinesOgrExporter(BaseOgrExporter):
             for field_name, field_type in six.iteritems(fields):
                 fname = LINE_FIELD_NAME_MAP.get(field_name, field_name)
                 if field_name == 'kcu_descr':
-                    value = ''
+                    raw_value = ''
                     try:
-                        value = str(kcu_dict[int(line_data['kcu'][i])])
+                        raw_value = str(kcu_dict[int(line_data['kcu'][i])])
                     except KeyError:
                         pass
                 elif field_name == 'node_a':
-                    value = TYPE_FUNC_MAP[field_type](node_a[i])
+                    raw_value = node_a[i]
                 elif field_name == 'node_b':
-                    value = TYPE_FUNC_MAP[field_type](node_b[i])
+                    raw_value = node_b[i]
                 else:
                     try:
                         raw_value = line_data[fname][i]
                     except IndexError:
-                        raw_value = '-9999'
-                    value = TYPE_FUNC_MAP[field_type](raw_value)
-                feature.SetField(str(field_name), value)
+                        raw_value = None
+
+                self.set_field(feature, field_name, field_type, raw_value)
 
             layer.CreateFeature(feature)
             feature.Destroy()