+++ HISTORY.rst
@@ -4,7 +4,15 @@ History
 1.1.11 (unreleased)
 -------------------
 
-- Nothing changed yet.
+- Handling of NaN, Inf and -Inf, -9999, and empty strings in exporters.
+  These get exported as NULL.
+
+- Fixed exporting string dtype fields (e.g. cont_type) in OGR exporter. For instance,
+  the string "b'something'" now is written as "something".
+
+- Skip the dummy element (with id=0) in all exporters.
+
+- Set the FID (feature ID) in the OGR (shapefile/geopackage/some geojson) exporters.
 
 
 1.1.10 (2021-11-01)